ml/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
matplotlib
scikit-image
opencv-python
tqdm
pillow
lpips
pytest
tensorboard
